feat(auth): auto-login to wiki externally authenticated user
GiantBomb Next (GBN) is responsible for authentication and cookies, while the
wiki will live on the same domain, under /wiki.

GBN will write a cookie on /wiki with a JWT containing the authenticated
user's information. Key to verify the token is available at /.well-known/jwks.json

We need a way to automatically login to Mediawiki when the user is
already authenticated to GBN. There's no existing extension to perform
this, so we need to write a custom extension to load our
SessionProvider. What needs to be done is described in this example
https://www.mediawiki.org/wiki/Manual:SessionManager_and_AuthManager/SessionProvider_examples